feat(admin):优化删除精选内容时逻辑
- 调整删除逻辑，确保数据库删除成功后再执行文件异步删除
- 增加删除失败时的错误响应返回
- 避免因数据库删除失败导致的文件残留问题

+++ api/api-admin/src/main/java/com/ligg/apiadmin/controller/AdminFeaturedController.java
@@ -130,9 +130,11 @@ public class AdminFeaturedController {
             return Response.error(BusinessStates.NOT_FOUND);
         }
         String imagePath = IMAGE_PATH + featured.getImagePath().replace("/api/image", "");
-        //异步删除
-        fileService.deleteFileAsync(imagePath);
-        featuredService.removeById(id);
-        return Response.success(BusinessStates.SUCCESS);
+        if (featuredService.removeById(id)) {
+            //异步删除
+            fileService.deleteFileAsync(imagePath);
+            return Response.success(BusinessStates.SUCCESS);
+        }
+        return Response.error(BusinessStates.INTERNAL_SERVER_ERROR);
     }
 }
